+++ CHANGELOG.md
@@ -5,6 +5,8 @@ v3.17.4 - 2018-04-??
 --------------------
 
 + Added core RTL language support
++ Fixed fatal error on student management tables resulting from deleted admin users who manually enrolled students
+
 
 ##### Template Updates